// src/lib/services/models/UserModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserModel = {
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    otherNames?: string | null;
    dateOfBirth?: string;
    password?: string | null;
    transactionPin?: number;
    securityQuestion?: string | null;
    securityAnswer?: string | null;
    phoneNumber?: string | null;
    referralId?: string | null;
    bvn?: string | null;
};